create legend component, remove use of createElement
