+++ packages/geoview-core/src/core/components/legend/legend-api.ts
@@ -1,16 +1,8 @@
 import { createElement } from 'react';
 import { LegendsLayerSet, api } from '@/app';
 import { LegendItem, TypeLegendItemProps } from './legend-item';
-import { List } from '@/ui';
-
-export interface TypeLegendProps {
-  layerIds: string[];
-  isRemoveable?: boolean;
-  canSetOpacity?: boolean;
-  expandAll?: boolean;
-  hideAll?: boolean;
-  canZoomTo?: boolean;
-}
+import { TypeLegendProps } from './types';
+import { Legend } from './legend';
 
 /**
  * API to manage legend component
@@ -38,25 +30,10 @@ export class LegendApi {
    *
    */
   createLegend = (props: TypeLegendProps) => {
-    const { layerIds, isRemoveable, canSetOpacity, expandAll, hideAll } = props;
-    api.event.emit({ handlerName: `${this.mapId}/$LegendsLayerSet$`, event: api.eventNames.GET_LEGENDS.TRIGGER });
-    const legendItems = layerIds.map((layerId) => {
-      const geoviewLayerInstance = api.maps[this.mapId].layer.geoviewLayers[layerId];
-      if (geoviewLayerInstance) {
-        return createElement(LegendItem, {
-          key: `layerKey-${layerId}`,
-          layerId,
-          geoviewLayerInstance,
-          isRemoveable,
-          canSetOpacity,
-          expandAll,
-          hideAll,
-          canZoomTo: true,
-        });
-      }
-      return null;
+    return createElement(Legend, {
+      ...props,
+      mapId: this.mapId,
     });
-    return createElement('div', {}, createElement(List, { sx: { width: '100%' } }, legendItems));
   };
 
   /**
