+++ packages/geoview-core/src/core/components/legend/types.ts
@@ -1,0 +1,8 @@
+export interface TypeLegendProps {
+  layerIds: string[];
+  isRemoveable?: boolean;
+  canSetOpacity?: boolean;
+  expandAll?: boolean;
+  hideAll?: boolean;
+  canZoomTo?: boolean;
+}
